refactor render prep, numeric parsing, and sensitive screening
